Testes unitarios e funcionais criados. Refatoracao do campo interestRate para int ao inves de double

+++ src/main/java/desafio/backend/emprestimo/dto/EmprestimoDTO.java
@@ -14,6 +14,6 @@ public record EmprestimoDTO(
         TipoEmprestimo type,
 
         @JsonProperty("interest_rate")
-        double interestRate) {
+        int interestRate) {
 
 }
